tutorials/backends/fastapi/strawberry/pyproject.toml
[project]
name = "strawberry-graphql-api"
version = "0.1.0"
description = "Supply Chain GraphQL API with Strawberry, FastAPI, and PostgreSQL"
requires-python = ">=3.11"

dependencies = [
    "strawberry-graphql==0.217.0",
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "asyncpg==0.29.0",
    "sqlalchemy==2.0.23",
    "greenlet==3.0.3",
    "pydantic==2.5.0",
    "python-dotenv==1.0.0",
    "httpx==0.25.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
